refactor(recipes): consolidate add-recipe flow into one sheet instance
The chooser, manual form, and URL import previously swapped between three
separate MobileSheet mounts mid-open, which remounted the drawer and broke
the open animation. A single sheet now switches title, content, and height
by mode (chooser/import at half height, manual form at full).
